requirements.txt
streamlit
transformers
torch
bert-extractive-summarizer
